+++ include/highfive/bits/H5FileDriver_misc.hpp
@@ -10,56 +10,13 @@
 #ifndef H5FILEDRIVER_MISC_HPP
 #define H5FILEDRIVER_MISC_HPP
 
-#include <H5Ppublic.h>
-
-#ifdef H5_HAVE_PARALLEL
-#include <H5FDmpi.h>
-#endif
-
 namespace HighFive {
 
-namespace {
-
-template <typename Comm, typename Info>
-class MPIOFileAccess
-{
-public:
-  MPIOFileAccess(Comm comm, Info info)
-      : _comm(comm)
-      , _info(info)
-  {}
-
-  void apply(const hid_t list) const {
-    if (H5Pset_fapl_mpio(list, _comm, _info) < 0) {
-        HDF5ErrMapper::ToException<FileException>(
-            "Unable to set-up MPIO Driver configuration");
-    }
-  }
-private:
-  Comm _comm;
-  Info _info;
-};
-
-}  //namespace
-
-template <typename Comm, typename Info>
-inline MPIOFileDriver::MPIOFileDriver(Comm comm, Info info) {
-    add(MPIOFileAccess<Comm, Info>(comm, info));
+#ifdef H5_HAVE_PARALLEL
+inline MPIOFileDriver::MPIOFileDriver(MPI_Comm comm, MPI_Info info) {
+    add(MPIOFileAccess(comm, info));
 }
-
-inline void FileVersionBounds::apply(const hid_t list) const {
-    if (H5Pset_libver_bounds(list, _low, _high) < 0) {
-        HDF5ErrMapper::ToException<PropertyException>(
-            "Error setting file version bounds");
-    }
-}
-
-inline void MetadataBlockSize::apply(const hid_t list) const {
-    if (H5Pset_meta_block_size(list, _size) < 0) {
-        HDF5ErrMapper::ToException<PropertyException>(
-            "Error setting metadata block size");
-    }
-}
+#endif
 
 } // namespace HighFive
